demogen/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
playwright-stealth==1.0.6
google-genai>=0.8.0
python-multipart==0.0.6
reportlab==4.0.7
pillow==10.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
